ProxyAccess/requirements.txt
allure-behave==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
astroid==2.11.5
attrs==22.2.0
behave==1.2.6
black==23.1.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
chromedriver-py==103.0.5060.24
click==8.0.3
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
PyHamcrest==2.0.4
PySocks==1.7.1
pylint==2.13.4
python-dateutil==2.8.2
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
uitestcore==10.3.1
urllib3==1.26.12
wsproto==1.2.0
git+https://github.com/nhsuk/axe-selenium-python.git
